refactor(desktop): consolidate delete dialog open coordination
