--- apps/desktop/src/renderer/react-query/workspaces/deferDeleteDialogOpen.ts
@@ -1,11 +1,0 @@
-/**
- * Defers opening the delete dialog to the next macrotask.
- * This avoids a focus race with Radix ContextMenu focus restoration.
- */
-export function deferDeleteDialogOpen(
-	setShowDeleteDialog: (show: boolean) => void,
-) {
-	return setTimeout(() => {
-		setShowDeleteDialog(true);
-	}, 0);
-}
