--- apps/desktop/src/renderer/screens/main/components/WorkspaceSidebar/WorkspaceListItem/context-menu-delete-dialog-coordinator.ts
@@ -1,28 +1,0 @@
-export interface AutoFocusEventLike {
-	preventDefault: () => void;
-}
-
-/**
- * Coordinates opening the delete dialog from a ContextMenu item selection.
- *
- * When "Close Worktree" is selected, we wait for ContextMenu close and then:
- * 1) prevent Radix auto-focus from returning to the trigger
- * 2) open the delete dialog
- */
-export function createContextMenuDeleteDialogCoordinator(
-	openDeleteDialog: () => void,
-) {
-	let shouldOpenDeleteDialog = false;
-
-	return {
-		requestOpenDeleteDialog() {
-			shouldOpenDeleteDialog = true;
-		},
-		handleCloseAutoFocus(event: AutoFocusEventLike) {
-			if (!shouldOpenDeleteDialog) return;
-			shouldOpenDeleteDialog = false;
-			event.preventDefault();
-			openDeleteDialog();
-		},
-	};
-}
